Внес правки, добавил добавление пользователя
вынеси логику из контроллера в сервис
не взаимодействуй с дао из контроллера напрямую
помечай транзакции для чтения как readOnly
транзакции в дао не нужны, они уже есть в сервисе
при редактировании метод контроллера должен принимать объект пользователя или его дто, но не поля по одному

+++ crudusers242/src/main/java/web/service/UserServiceImpl.java
@@ -2,74 +2,98 @@ package web.service;
 
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.transaction.annotation.Transactional;
+
 import web.dao.UserDao;
 import web.dao.RoleDAO;
-import web.dao.UserDao;
+import web.exceptions.UserNotFoundException;
 import web.model.Role;
 import web.model.User;
-import org.springframework.stereotype.Service;
 
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 @Service
-@Transactional
 public class UserServiceImpl implements UserService, UserDetailsService {
     private UserDao userDAO;
     private RoleDAO roleDAO;
 
+
     @Autowired
     PasswordEncoder bCryptPasswordEncoder;
 
-    @Transactional
+    
     @Autowired
     public void setUserDAO(UserDao userDAO) {
         this.userDAO = userDAO;
     }
 
-    @Transactional
+    
     @Autowired
     public void setRoleDAO(RoleDAO roleDAO) {
         this.roleDAO = roleDAO;
     }
 
-    @Transactional
+    
     @Override
     public List<User> allUsers() {
         return userDAO.allUsers();
     }
 
-    @Transactional
+    
     @Override
-    public void save(User user) {
-        user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
+    public void save(User user, Set<Role> roles) {
+        user.setRoles(roles);
+        setPassword(user);
         userDAO.save(user);
     }
 
-    @Transactional
+    
     @Override
-    public void delete(User user) {
+    public void delete(Long id) {
+        User user = userDAO.getById(id);
+        if (user == null) throw new UserNotFoundException();
         userDAO.delete(user);
     }
 
-    @Transactional
+    
     @Override
     public User getById(Long id) {
         return userDAO.getById(id);
     }
 
-    @Transactional
+    
     @Override
     public UserDetails loadUserByUsername(String username) {
-        User user = userDAO.getUserByName(username);
-        return user;
+        return userDAO.getUserByName(username);
+    }
+
+
+    @Override
+    public void edit(User user, List<String> roles) {
+        final User userDB = userDAO.getById(user.getId());
+        if (userDB == null) throw new UserNotFoundException();
+        Set<Role> Setroles = new HashSet<>();
+        for (String st : roles) {
+            if (st.equals("ADMIN")) {
+                Role role_admin = roleDAO.createRoleIfNotFound("ADMIN", 1L);
+                Setroles.add(role_admin);
+            }
+            if (st.equals("USER")) {
+                Role role_user = roleDAO.createRoleIfNotFound("USER", 2L);
+                Setroles.add(role_user);
+            }
+        }
+        user.setRoles(Setroles);
+        setPassword(user);
+        userDAO.save(user);
+    }
+
+    private void setPassword(User user) {
+        if (user.getPassword() != null) {
+            user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
+        }
     }
 }
